+++ addOns/commonlib/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- An HTTP date parser/formatter.
+
 ### Fixed
 - Take into account the timezone when checking if a cookie is expired (Issue 6550).